feat: add casos de teste para o ClusteredMultiselect
